pyproject.toml
[tool.poetry]
name = "mini-parsec"
version = "0.1.0"
description = "Test SSE"
authors = ["Maximilien de Dinechin <m.dinechin@protonmail.com>"]
readme = "README.md"
packages = [{include = "miniparsec"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.6"
rich = "^13.4.1"
psycopg = {extras = ["binary", "pool"], version = "^3.1.9"}
tqdm = "^4.65.0"
requests = "^2.31.0"
pynacl = "^1.5.0"
watchdog = "^3.0.0"
unidecode = "^1.3.6"
textract = "^1.6.5"
pycountry = "^22.3.5"
pystemmer = "^2.2.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ruff = "^0.0.270"
isort = "^5.12.0"
snakeviz = "^2.2.0"
pyright = "^1.1.311"
mypy = "^1.3.0"
types-requests = "^2.31.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
